Made change to double nan instead of long double.

+++ agent/globals.hpp
@@ -38,7 +38,7 @@
 
 #ifdef _WINDOWS
 #define ISNAN(x) _isnan(x)
-#define NAN numeric_limits<long double>::quiet_NaN()
+#define NAN numeric_limits<double>::quiet_NaN()
 typedef unsigned __int64 uint64_t;
 #define strtoull _strtoui64
 #else
